+++ runner/android_junit_runner/javatests/androidx/test/internal/runner/RunnerArgsTest.java
@@ -755,5 +755,10 @@ public class RunnerArgsTest {
     public OutputStream openInternalOutputFile(String pathname) throws IOException {
       throw new UnsupportedOperationException();
     }
+
+    @Override
+    public boolean isTestStorageFilePath(String pathname) {
+      return false;
+    }
   }
 }