+++ ext/junit/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 **Bug Fixes**
 
+* Use PlatformTestStorage instead of TestStorage in DeleteFilesRule 
+
 **New Features**
 
 **Breaking Changes**