+++ runner/monitor/CHANGELOG.md
@@ -12,6 +12,7 @@
 
 **API Changes**
 * Make DeviceController an public API from ExperimentalTestApi
+* Upstream TestStorage.isTestStoragePath to PlatformTestStorage
 
 **Breaking API Changes**